Split TableNode from ast_builder/mod.rs, add ast_builder/table.rs
