+++ core/src/ast_builder/table.rs
@@ -1,0 +1,16 @@
+use super::{DeleteNode, SelectNode};
+
+#[derive(Clone)]
+pub struct TableNode {
+    pub table_name: String,
+}
+
+impl TableNode {
+    pub fn select(self) -> SelectNode {
+        SelectNode::new(self.table_name)
+    }
+
+    pub fn delete(self) -> DeleteNode {
+        DeleteNode::new(self.table_name)
+    }
+}
